requirements.txt
certifi==2019.3.9
chardet==3.0.4
data==0.4
decorator==4.4.0
Django==2.2
funcsigs==1.0.2
future==0.17.1
idna==2.8
latex==0.7.0
numpy==1.16.2
ordered-set==3.1
PyLaTeX==1.3.0
pytz==2019.1
requests==2.21.0
scipy==1.2.1
shutilwhich==1.1.0
six==1.12.0
sqlparse==0.3.0
stripe==2.25.0
tempdir==0.7.1
urllib3==1.24.2
django-stdimage==5.0.0
